requirements.txt
fastapi
uvicorn
requests
python-multipart
pydantic
openai-whisper
torch
numpy
